requirements.txt
# requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Algerian Document Classifier — GPU (4-bit quantized) version
# Install with: pip install -r requirements.txt
#
# For Google Colab: run the Colab setup cells instead (see README).
# For local GPU: verify your CUDA version first:
#   nvidia-smi | head -1
# Then replace "cu121" below with your version (cu118, cu121, cu124, etc.)
# ─────────────────────────────────────────────────────────────────────────────

# Web framework
fastapi
uvicorn[standard]
python-multipart

# Image handling
Pillow

# PyTorch — CUDA build
# Replace cu121 with your actual CUDA version (e.g. cu118 or cu124)
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision

# HuggingFace model loading (install from source for Qwen2.5-VL support)
git+https://github.com/huggingface/transformers.git
accelerate

# 4-bit quantization — REQUIRED for this version
bitsandbytes>=0.43.0

# Qwen-VL vision utilities
qwen-vl-utils

# HTTP client for test_api.py
requests
